fix : reviewCount Integer 형으로 변경

+++ concertseat/src/main/java/site/concertseat/domain/review/service/ReviewService.java
@@ -20,7 +20,7 @@ public interface ReviewService {
 
     ImageUploadRes uploadImage(Member member, List<MultipartFile> file) throws IOException;
 
-    Long getTotalReviewCount();
+    Integer getTotalReviewCount();
 
     MyReviewStadiumListRes findStadium(Member member);
